// src/types.tsx
export interface IPuppy {
    id: string
    breed: string
    name: string
    image: string
    birthDate: string
}
export type Puppy = {
    id: string;
    breed: string;
    name: string;
    image: string;
    birthDate: string;
  }
  export interface IPerson{
    id: number,
    developerName: string,
    qualification: string,
    expertise: string,
    imageUrl: string
  }